Added some comments
- Added some comments in createserv-server.js and register.js :)
- Moved useful fs-related functions to new fsPlus.js file.
