+++ private/register.js
@@ -1,5 +1,6 @@
 var path = require('path');
 var mods = require('./getProps.js');
+var fsExt = require('./fsPlus.js');
 var express = mods.express;
 var app = mods.app;
 var http = mods.http;
@@ -31,6 +32,7 @@ io.on('connection', function(socket){
 					return printError(err, 1);
 				}
 				
+				// Hash password
 				bcrypt.hash(data.pass, salt, function(err, hash) { 
 					if(err) {
 						return printError(err, 2);
@@ -43,11 +45,15 @@ io.on('connection', function(socket){
 						
 						var files = 0;
 						var currentFile = 0;
+						
+						// Count files
 						li.forEach(function(file) {
 							files += 1;
 						});
 						
 						if(files > 0) {
+							
+							// Directory isn't empty, search the database to check if the user already exists
 							li.forEach(function(file) {
 								var dat = fs.readFileSync("users/" + file, 'utf8');
 								values = dat.split("\n");
@@ -66,17 +72,21 @@ io.on('connection', function(socket){
 							return;
 						}
 						
+						// User doesn't exist yet, register new user
 						fs.readFile("users/user.txt", 'utf8', function(error, dat) {
 							if(error) {
 								return printError(error, 5);
 							}
 							
 							values = dat.split("\n");
+							
+							// Add email & password to user file
 							fs.writeFile("users/" + values[0].toString() + ".txt", data.email + "\n" + hash, function(err, data) {
 								if(err) {
 									return printError(err, 6);
 								}
 								
+								// Make sure next user registered doesn't get the same user id
 								fs.writeFile("users/user.txt", Number(values[0]) + 1, function(err, data) {
 									if(err) {
 										return printError(err, 7);
